requirements.txt
torch>=1.6.0+cu101
PyYAML>=5.3.1
numpy>=1.19.2
nibabel>=3.1.1
scipy>=1.5.3
scikit-image>=0.17.2
torchvision>=0.7.0+cu101
https://github.com/ai-med/squeeze_and_excitation/releases/download/v1.0/squeeze_and_excitation-1.0-py2.py3-none-any.whl
tqdm>=4.50.2
pandas>=1.1.5